REFACTOR: Refactor map javascript
